Move pending withdrawals to tunnel tabs
