+++ webapp/app/[locale]/layout.tsx
@@ -2,7 +2,6 @@ import 'styles/globals.css'
 import '@rainbow-me/rainbowkit/styles.css'
 import 'react-loading-skeleton/dist/skeleton.css'
 
-import { TunnelHistoryProvider } from 'app/context/tunnelHistoryContext'
 import { locales, type Locale } from 'app/i18n'
 import { AppScreen } from 'components/appScreen'
 import { ErrorBoundary } from 'components/errorBoundary'
@@ -42,16 +41,14 @@ export default async function RootLayout({
         <NextIntlClientProvider locale={locale} messages={messages}>
           <RecaptchaContext>
             <WalletsContext locale={locale}>
-              <TunnelHistoryProvider>
-                <div className="flex h-dvh flex-nowrap justify-stretch">
-                  <div className="hidden w-1/4 max-w-56 md:block">
-                    <Navbar />
-                  </div>
-                  <AppScreen>
-                    <ErrorBoundary>{children}</ErrorBoundary>
-                  </AppScreen>
+              <div className="flex h-dvh flex-nowrap justify-stretch">
+                <div className="hidden w-1/4 max-w-56 md:block">
+                  <Navbar />
                 </div>
-              </TunnelHistoryProvider>
+                <AppScreen>
+                  <ErrorBoundary>{children}</ErrorBoundary>
+                </AppScreen>
+              </div>
             </WalletsContext>
           </RecaptchaContext>
         </NextIntlClientProvider>
