+++ webapp/app/[locale]/navbar/navData.tsx
@@ -7,20 +7,11 @@ import { ExplorerIcon } from 'components/icons/explorerIcon'
 import { FiletextIcon } from 'components/icons/filetextIcon'
 import { GraduateCapIcon } from 'components/icons/graduateCapIcon'
 import { TunnelIcon } from 'components/icons/tunnelIcon'
-import dynamic from 'next/dynamic'
 
 import { NavItemData } from './_components/navItems'
 
-const ActionableWithdrawals = dynamic(
-  () =>
-    import('./_components/actionableWithdrawals').then(
-      mod => mod.ActionableWithdrawals,
-    ),
-  { ssr: false },
-)
 export const navItems: NavItemData[] = [
   {
-    alertComponent: () => <ActionableWithdrawals />,
     href: '/tunnel',
     icon: TunnelIcon,
     id: 'tunnel',
