+++ webapp/app/[locale]/tunnel/_components/actionableWithdrawals.tsx
@@ -16,7 +16,7 @@ export const ActionableWithdrawals = function () {
   }
 
   return (
-    <div className="-mt-1 flex aspect-square min-w-6 items-center justify-center rounded-full bg-orange-950 p-1 text-center text-xs text-white">
+    <div className="flex aspect-square min-w-6 items-center justify-center rounded-full bg-orange-950 p-1 text-center text-xs text-white">
       {actionableWithdrawals}
     </div>
   )
